+++ packages/dart_http_native_client/lib/src/generated_bindings.dart
@@ -108,6 +108,23 @@ external int dart_http_native_client_websocket_send_binary(
   int length,
 );
 
+@ffi.Native<
+  ffi.Int64 Function(
+    ffi.Int64,
+    ffi.Int64,
+    ffi.Pointer<ffi.Uint8>,
+    ffi.IntPtr,
+    ffi.Pointer<ffi.Void>,
+  )
+>()
+external int dart_http_native_client_websocket_send_binary_native_prefixed(
+  int client_id,
+  int socket_id,
+  ffi.Pointer<ffi.Uint8> prefix,
+  int prefix_length,
+  ffi.Pointer<ffi.Void> native_buffer,
+);
+
 @ffi.Native<ffi.Int64 Function(ffi.Int64, ffi.Int64, ffi.Int32, ffi.Pointer<ffi.Char>)>()
 external int dart_http_native_client_websocket_close(
   int client_id,